# docs/swagger.yaml
basePath: /api
definitions:
  models.Attendance:
    properties:
      checked_in_at:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      session:
        allOf:
        - $ref: '#/definitions/models.AttendanceSession'
        description: Relationships
      session_id:
        type: integer
      student_name:
        type: string
      updated_at:
        type: string
      work_unit:
        type: string
      work_unit_address:
        type: string
    type: object
  models.AttendanceSession:
    properties:
      attendances:
        items:
          $ref: '#/definitions/models.Attendance'
        type: array
      class:
        $ref: '#/definitions/models.Class'
      class_id:
        type: integer
      created_at:
        type: string
      event:
        allOf:
        - $ref: '#/definitions/models.Event'
        description: Relationships
      event_id:
        type: integer
      id:
        type: integer
      session_date:
        type: string
      teacher:
        $ref: '#/definitions/models.Teacher'
      teacher_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Class:
    properties:
      class_code:
        type: string
      class_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      sessions:
        items:
          $ref: '#/definitions/models.AttendanceSession'
        type: array
      students:
        description: Relationships
        items:
          $ref: '#/definitions/models.Student'
        type: array
      updated_at:
        type: string
    type: object
  models.CreateAttendanceRequest:
    properties:
      email:
        example: student@example.com
        type: string
      phone:
        example: "0123456789"
        type: string
      session_id:
        example: 1
        type: integer
      student_name:
        example: Nguyen Van A
        type: string
      work_unit:
        example: Công ty ABC
        type: string
      work_unit_address:
        example: 123 Đường ABC, Quận 1, TP.HCM
        type: string
    required:
    - email
    - phone
    - session_id
    - student_name
    - work_unit
    - work_unit_address
    type: object
  models.CreateAttendanceSessionRequest:
    properties:
      class_id:
        example: 1
        type: integer
      event_id:
        example: 1
        type: integer
      session_date:
        example: "2025-08-20T08:31:46.121Z"
        type: string
      teacher_id:
        example: "1"
        type: string
    type: object
  models.CreateClassRequest:
    properties:
      class_code:
        example: LOP001
        type: string
      class_name:
        example: Lớp Khoa học máy tính K65
        type: string
    required:
    - class_code
    - class_name
    type: object
...
  models.Event:
    properties:
      created_at:
        type: string
      description:
        type: string
      event_name:
        type: string
      id:
        type: integer
      sessions:
        description: Relationships
        items:
          $ref: '#/definitions/models.AttendanceSession'
        type: array
      start_date:
        type: string
      updated_at:
        type: string
    type: object
  models.Student:
    properties:
      class:
        allOf:
        - $ref: '#/definitions/models.Class'
        description: Relationships
      class_id:
        type: integer
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      student_code:
        type: string
      student_name:
        type: string
      updated_at:
        type: string
      work_unit:
        type: string
    type: object
  models.Teacher:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      sessions:
        description: Relationships
        items:
          $ref: '#/definitions/models.AttendanceSession'
        type: array
      teacher_code:
        type: string
      teacher_name:
        type: string
      updated_at:
        type: string
      work_unit:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple student management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Student Management API
  version: "1.0"
paths:
  /attendance-sessions:
    get:
      description: Get all attendance sessions with class and teacher information
      parameters:
      - description: Filter by Event ID
        in: query
        name: event_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AttendanceSession'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all attendance sessions
      tags:
      - attendance-sessions
    post:
      consumes:
      - application/json
      description: Create a new attendance session in the database
      parameters:
      - description: Attendance session data
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/models.CreateAttendanceSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AttendanceSession'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new attendance session
      tags:
      - attendance-sessions
  /attendance-sessions/{id}:
    get:
      description: Get a specific attendance session with class, teacher, and attendances
      parameters:
      - description: Attendance Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttendanceSession'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get attendance session by ID
      tags:
      - attendance-sessions
  /attendances:
    get:
      description: Get all attendance records with session information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attendance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all attendances
      tags:
      - attendances
    post:
      consumes:
      - application/json
      description: Create a new attendance record
      parameters:
      - description: Attendance data
        in: body
        name: attendance
        required: true
        schema:
          $ref: '#/definitions/models.CreateAttendanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Attendance'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new attendance
      tags:
      - attendances
  /attendances/{id}:
    get:
      description: Get a specific attendance record with session information
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Attendance'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get attendance by ID
      tags:
      - attendances
  /classes:
    get:
      description: Get all classes from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Class'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all classes
      tags:
      - classes
    post:
      consumes:
      - application/json
      description: Create a new class in the database
      parameters:
      - description: Class data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.CreateClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Class'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new class
      tags:
      - classes
  /classes/{id}:
    get:
      description: Get a specific class with its students
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Class'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get class by ID
      tags:
      - classes
...
  /sessions/{sessionId}/attendances:
    get:
      description: Get all attendance records for a specific session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attendance'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get attendances by session ID
      tags:
      - attendances
...
swagger: "2.0"